Fixed some compiler warnings for DESKTOP build

+++ Software/src/engine/FrameBuffer.cpp
@@ -65,7 +65,7 @@ void FrameBuffer::drawHorizontalLine(int x1, int y, int x2, uint16_t color) {
 			x >= 0
 			&& x < WIDTH
 			&& dest >= 0
-			&& dest < FRAMEBUFFER_SIZE
+			&& dest < (int32_t)FRAMEBUFFER_SIZE
 		) {
 			frame[dest]=color;
 		}
@@ -83,7 +83,7 @@ void FrameBuffer::drawVerticalLine(int x, int y1, int y2, uint16_t color) {
 			y >= 0
 			&& y < HEIGHT
 			&& dest >= 0
-			&& dest < FRAMEBUFFER_SIZE
+			&& dest < (int32_t)FRAMEBUFFER_SIZE
 		) {
 			frame[dest] = color;
 		}
